Bug 5734 - Fixed metric conversion on the vehicle/driver speed detail page

+++ dao/src/main/java/com/inthinc/pro/model/AggressiveDrivingEvent.java
@@ -81,7 +81,7 @@ public class AggressiveDrivingEvent extends Event
             speed = this.getSpeed();
         
         if(measurementType.equals(MeasurementType.METRIC))
-            speed = MeasurementConversionUtil.fromMPHtoKPH(speed);
+            speed = MeasurementConversionUtil.fromMPHtoKPH(speed).intValue();
         return MessageFormat.format(formatStr, new Object[] {speed, mphString});
     }
